Shared getter across submodules
