+++ src/zeroband/inference/pipeline.py
@@ -13,8 +13,8 @@ from vllm.executor.mp_distributed_executor import MultiprocessingDistributedExec
 from vllm.model_executor.layers.sampler import SamplerOutput
 
 from zeroband.inference.config import PipelineParallelConfig
-from zeroband.inference.logger import get_logger
 from zeroband.inference.utils import rgetattr
+from zeroband.utils.logger import get_logger
 
 
 def serialize_tensors(tensor_dict: dict[str, torch.Tensor]) -> bytes:
